rework variants/pxb-layout
