+++ components/src/core/drawer/drawer-layout/drawer-layout.component.scss
@@ -1,35 +1,9 @@
 .pxb-drawer-layout {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-
-    .pxb-drawer-layout-drawer-wrapper {
-        display: flex;
-        position: fixed;
-        height: 100%;
-        align-items: stretch;
-        overflow-y: scroll;
-    }
-
-    .pxb-drawer-layout-content-wrapper {
-        transition: padding 175ms cubic-bezier(0.4, 0, 0.2, 1);
+    .mat-drawer-side {
+        border-right: none;
     }
 }
 
-.pxb-drawer-layout-content-overlay::after {
-    display: flex;
-    content: '';
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: rgba(0, 0, 0, 0.5);
-    z-index: 1;
-}
-
-@media screen and (max-width: 600px) {
-    .pxb-drawer-layout-content-wrapper {
-        padding-left: 0px !important;
-    }
+mat-sidenav.pxb-drawer-closed.pxb-drawer-persistent {
+    width: 52px;
 }
