+++ demos/showcase/src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Component, ViewEncapsulation } from '@angular/core';
+import { Component, ViewEncapsulation, Input } from '@angular/core';
 import { MatIconRegistry } from '@angular/material/icon';
 import { DomSanitizer } from '@angular/platform-browser';
 import * as PXBColors from '@pxblue/colors';
@@ -14,6 +14,8 @@ const iconSet = require('@pxblue/icons-svg/icons.svg');
 export class AppComponent {
     drawerBodyTestItems;
     colors: Record<string, any>;
+    drawerOpen = true;
+    variant = 'persistent';
     constructor(private readonly matIconRegistry: MatIconRegistry, private readonly domSanitizer: DomSanitizer) {
         this.colors = PXBColors;
         this.matIconRegistry.addSvgIconSetInNamespace(
@@ -31,4 +33,16 @@ export class AppComponent {
         // eslint-disable-next-line
         console.log('drawer header button clicked...');
     }
+
+    toggleDrawer(): void {
+        if (this.variant !== 'permanent') {
+            this.drawerOpen = !this.drawerOpen;
+        }
+    }
+
+    setVariant(str: string): void {
+        this.drawerOpen = true;
+        this.variant = str;
+    }
+
 }
